Move `RegistrationResult` to the `user.registeruser` package
