+++ app/src/main/java/org/simple/clinic/user/registeruser/RegistrationResult.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.registration
+package org.simple.clinic.user.registeruser
 
 sealed class RegistrationResult {
